Move shrimpy/dynatrack/preprocessing.py to shrimpy/preprocessing.py
Pure move: the module's contents are byte-for-byte identical to before, so
this commit is a 100%-similarity rename. Only the three references to it are
repointed (the worker's import, two docstring cross-references in tracking.py,
and the package-layout table in the DynaTrack README).

The deskew -> phase -> virtual-staining pipeline is general label-free
reconstruction rather than something specific to position tracking, and smart
FOV selection needs exactly the same reconstruction for its pre-scan. Reaching
into shrimpy.dynatrack from shrimpy.fov_selection would couple two features
for no reason, and copying the module would leave two implementations to drift
apart.

Decoupling it from DynaTrackConfig is the next commit; this one moves the file
and nothing else, so the rename is reviewable on its own.
